refactor: move code to a dedicated folder
