+++ src/app/shared/easterEgg/konamiCode.ts
